requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
auth0-python==4.9.0
backoff==2.2.1
bcrypt==4.3.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==1.0.0
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
crewai==0.108.0
cryptography==44.0.2
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
docstring-parser==0.16
durationpy==0.9
et-xmlfile==2.0.0
executing==2.2.0
fastapi==0.115.9
filelock==3.18.0
Flask==3.1.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.3.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
groq==0.22.0
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.30.1
humanfriendly==10.0
idna==3.10
importlib-metadata==8.6.1
importlib-resources==6.5.2
instructor==1.7.9
ipython==9.0.2
ipython-pygments-lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
json5==0.12.0
json-repair==0.40.0
jsonpatch==1.33
jsonpickle==4.0.5
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==32.0.1
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-groq==0.3.2
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.8
langsmith==0.3.24
litellm==1.60.2
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.3.2
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.4
oauthlib==3.2.2
onnxruntime==1.21.0
openai==1.70.0
openpyxl==3.1.5
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
orjson==3.10.16
overrides==7.7.0
packaging==24.2
parso==0.8.4
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.1.0
posthog==3.23.0
prompt-toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pyvis==0.3.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.2
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.45.3
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
tomli_w==1.2.0
torch==2.6.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.3
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uv==0.6.12
uvicorn==0.34.0
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0